feat: implement Android experimental backend with CommandQueue API
Custom TextureView renderer with Choreographer render loop, CommandQueue
polling infrastructure, ViewModel resolution and property validation,
and example Compose test activities for manual testing.

+++ android/src/experimental/java/com/margelo/nitro/rive/HybridViewModelColorProperty.kt
@@ -28,7 +28,7 @@ class HybridViewModelColorProperty(
       }
     }
     set(value) {
-      instance.setColor(path, value.toInt())
+      instance.setColor(path, value.toLong().toInt())
     }
 
   override fun addListener(onChanged: (value: Double) -> Unit): () -> Unit {
